Updated pending events. Working on buttons
